requirements.txt
requests
python-dotenv
openai
exa_py
beautifulsoup4
